FIXED: When the window width is wide, the responsive navbar panel is not collapsed. Then I see the Page Contents, and I can hover a mouse over an item in that list, and I get a hand cursor, and I can click the link, and it jumps to the corresponding header in the main panel.
When the responsive navbar collapses, then I see a hamburger menu called "Contents".
When I click that, the Page Content uncollapses and shows, but when the mouse hovers over an item, I do not get a hand cursor, and when I click an item, it does not jump to the corresponding header in the main panel, but rather the navbar collapses.

But the Hamburger Content tag disappears after clicking.